restructure file test
