+++ src/tests/component/Nav.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { Navbar } from 'react-bootstrap';
-import Nav from '../components/Nav';
+import Nav from '../../components/Nav';
 
 describe('Nav', () => {
   it('should render a <Nav />', () => {
